Made get_graph a bit less memory hungry, by using a float[][] instead of double[][] Not printing warnings about missing atoms anymore

+++ proteinstructure/Box.java
@@ -25,34 +25,34 @@ public class Box {
 		j_pointsInBox.put(serial, point);
 	}
 
-	public void getDistancesWithinBox(double[][] distMatrix, boolean directed){ //we pass a reference to the distMatrix that we alter within this method
+	public void getDistancesWithinBox(float[][] distMatrix, boolean directed){ //we pass a reference to the distMatrix that we alter within this method
 		for (int i_serial:i_pointsInBox.keySet()) {
 			for (int j_serial:j_pointsInBox.keySet()) {
 				if (!directed) {
 					if (j_serial>i_serial) { 
 						// this only works if previously we have made sure that atom serials are sequential from 0 to MAXATOMSERIAL
-						distMatrix[i_serial][j_serial] = i_pointsInBox.get(i_serial).distance(j_pointsInBox.get(j_serial));
+						distMatrix[i_serial][j_serial] = (float)i_pointsInBox.get(i_serial).distance(j_pointsInBox.get(j_serial));
 					} 
 				} else {
-					distMatrix[i_serial][j_serial] = i_pointsInBox.get(i_serial).distance(j_pointsInBox.get(j_serial));
+					distMatrix[i_serial][j_serial] = (float)i_pointsInBox.get(i_serial).distance(j_pointsInBox.get(j_serial));
 				}
 			}
 		}
 
 	}
 	
-	public void getDistancesToNeighborBox(Box nbBox ,double[][] distMatrix, boolean directed){
+	public void getDistancesToNeighborBox(Box nbBox ,float[][] distMatrix, boolean directed){
 		for (int i_serial:i_pointsInBox.keySet()){
 			for (int j_serial:nbBox.j_pointsInBox.keySet()){
 				if (!directed) {
 					if (j_serial>i_serial) {
 						// this only works if previously we have made sure that atom serials are sequential from 0 to MAXATOMSERIAL
-						if (distMatrix[i_serial][j_serial]==0.0){ // i.e. if we haven't passed through this cell yet
-							distMatrix[i_serial][j_serial] = i_pointsInBox.get(i_serial).distance(nbBox.j_pointsInBox.get(j_serial));
+						if (distMatrix[i_serial][j_serial]==0.0f){ // i.e. if we haven't passed through this cell yet
+							distMatrix[i_serial][j_serial] = (float)i_pointsInBox.get(i_serial).distance(nbBox.j_pointsInBox.get(j_serial));
 						}
 					}
 				} else {
-					distMatrix[i_serial][j_serial] = i_pointsInBox.get(i_serial).distance(nbBox.j_pointsInBox.get(j_serial));
+					distMatrix[i_serial][j_serial] = (float)i_pointsInBox.get(i_serial).distance(nbBox.j_pointsInBox.get(j_serial));
 				}
 			}
 		}
